Add a documentation site (#13)
This PR adds the skeleton for a GitHub pages site for Dusk. As part of
the site C and C++ API documentation is generated from `dawn.json`. This
is then combined with api information in the `src/api.yaml` file. The
documentation is then generated into the `docs/api/{c,cpp}.md` files to
be served.

The `api.yaml` file is no where near complete at this point, there are a
lot of warnings emitted about missing information, but should have
enough there to give an idea of how things would look.

The idea is, in the future, to write some text around each of the
examples as more of a tutorial then exists in the current READMEs. This
text, along with the READMEs would then be able to link to the generated
C and C++ APIs as needed.

Issue #10

Co-authored-by: Ben Clayton <bclayton@google.com>