Make a `common` folder.
The initial idea was that each example would be standalone with a copy
of all the needed code. This turned out to be a maintenance issue,
especially with the webgpu.h in flux. The amount of duplicated code
fixes made it untenable. Switch to a `common` folder where code used in
multiple examples will live.