Update cube movement.
Update the 4th and 5th examples to make the cube look nicer when moved.
The 5th is changed so it just sends simple information through the
uniform and calculates the cube matrix in the shader. The example is
updated to 1024*1024 cubes instead of 32*32 since it is no longer
limited to uniform buffer size.